+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+* Add connect button in administration user profile
 * Add configuration in administration
 * Add option to revoke connections in personal user profile
 * Add custom login grant